+++ camdrv/isp2.5/sw_isp/src/sprd_realtimebokeh.c
@@ -388,7 +388,7 @@ static void* misc_process(void* handle)
 		&aem_addr , &aem_size);
 #endif
 //	endTime("sw_isp_process cost");
-	SWISP_LOGI("----yzl add systemTime sw_isp_process cost time:%zd ms ,ret:%d\n" , (systemTime(CLOCK_MONOTONIC) - time)/1000000 , ret);
+	SWISP_LOGI("----yzl add systemTime sw_isp_process cost time:%lld ms ,ret:%d\n" , (long long)(systemTime(CLOCK_MONOTONIC) - time)/1000000 , ret);
 	if(ret != 0)
 	{
 		SWISP_LOGE("sw_isp_process failed ret:%d" , ret);
@@ -481,8 +481,8 @@ static void* misc_process(void* handle)
 	distance = DISTANCE_OK;
 //	memset(phandle->depth_weight_ptr[(phandle->depth_weight_valid_index+1)%2] , 0 , 960*720);
 //	memset((void*)phandle->depth_weight_ptr[(phandle->depth_weight_valid_index+1)%2] , 0 , 960*720);
-	SWISP_LOGI("-------micro process after Depth calculate use output index:%d , depthrun costtime:%zd ms, distance:%d , ret:%d\n" , (phandle->depth_weight_valid_index+1)%2 ,
-		   (systemTime(CLOCK_MONOTONIC)-time)/1000000 , distance , ret);
+	SWISP_LOGI("-------micro process after Depth calculate use output index:%d , depthrun costtime:%lld ms, distance:%d , ret:%d\n" , (phandle->depth_weight_valid_index+1)%2 ,
+		   (long long)(systemTime(CLOCK_MONOTONIC)-time)/1000000 , distance , ret);
 	if(ret != 0)
 	{
 		SWISP_LOGW("sprd_depth_Run return error:%d" , ret);
@@ -523,7 +523,7 @@ static void* misc_process(void* handle)
 	set_misc_flag(phandle , 0);
 	event_callback(phandle , SOFT_BUFFER_RELEASE , &phandle->misc_arg.raw ,&phandle->misc_arg.m_yuv_pre , 0,0);
 	event_callback(phandle , SOFT_IRQ_AEM_STATIS , NULL ,NULL , aem_addr , aem_size);
-	SWISP_LOGI("misc_process cost time:%zd ms" , (systemTime(CLOCK_MONOTONIC) - timestart)/1000000);
+	SWISP_LOGI("misc_process cost time:%lld ms" , (long long)(systemTime(CLOCK_MONOTONIC) - timestart)/1000000);
 	return NULL;
 }
 
@@ -731,7 +731,7 @@ __attribute__ ((visibility("default"))) int sprd_realtimebokeh_process(void* han
 		ret = iBokehBlurImage(phandle->bokeh_handle , (void*)isp_param->m_yuv_pre.graphicbuffer ,
 				      (void*)isp_param->m_yuv_bokeh.graphicbuffer);
 #endif
-		SWISP_LOGI("bokeh cost time:%zd ms ret:%d\n" , (systemTime(CLOCK_MONOTONIC)-time)/1000000 , ret);
+		SWISP_LOGI("bokeh cost time:%lld ms ret:%d\n" , (long long)(systemTime(CLOCK_MONOTONIC)-time)/1000000 , ret);
 		if(ret != 0)
 		{
 			SWISP_LOGE("iBokehBlurImage failed ret:%d" , ret);
@@ -766,7 +766,7 @@ __attribute__ ((visibility("default"))) int sprd_realtimebokeh_process(void* han
 				break;
 			}
 		}
-		SWISP_LOGI("bokeh use memcpy cost time:%zd ms\n" , (systemTime(CLOCK_MONOTONIC)-time)/1000000);
+		SWISP_LOGI("bokeh use memcpy cost time:%lld ms\n" , (long long)(systemTime(CLOCK_MONOTONIC)-time)/1000000);
 	}
 
 	{
